pokracovani velkeho uklidu na uzivatelpresenteru
