+++ app/templates/Uzivatel/list.latte
@@ -3,8 +3,8 @@
 <a href="{link Ap:list id=>$ap->Oblast->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-arrow-up"></i> Zobrazit oblast</a>
 <a href="{link Ap:show id=>$ap->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-eye-open"></i> Zobrazit podrobnosti AP</a>
 <a href="{link Wewimo:show id=>$ap->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-signal"></i> Wewimo</a>
-<a href="{link Uzivatel:emailall id=>$ap->id}" class="btn btn-default btn-sm" role="button"><b>@</b> Odeslat E-mail všem z AP</a>
-<a href="{link Uzivatel:smsall id=>$ap->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-phone"></i> Odeslat SMS všem z AP</a>
+<a href="{link UzivatelMailSms:emailall id=>$ap->id}" class="btn btn-default btn-sm" role="button"><b>@</b> Odeslat E-mail všem z AP</a>
+<a href="{link UzivatelMailSms:smsall id=>$ap->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-phone"></i> Odeslat SMS všem z AP</a>
 
 &nbsp;&nbsp;
 {if $canViewOrEdit}
